// src/types/api.ts
export type RequestStatus = 
  | 'Новая' 
  | 'Ожидает' 
  | 'Ожидает принятия' 
  | 'Принял' 
  | 'В пути' 
  | 'В работе' 
  | 'Готово' 
  | 'Отказ' 
  | 'Модерн' 
  | 'НеЗаказ' 
  | 'Перезвонить' 
  | 'ТНО';